Some refactoring to allow reduce operations
The compiler had to be refactored in order to make some room for new
operations like "reduce", "map" and "filter". In order to do that, it
was necessary to create an abstract concept of "operation" replacing the
old "iterator" element. From now on, all these operations will be
represented by "Operation" element in the intermediate representation
level.