Preparing support for map and filter operations
First refactoring to allow support for map and filter operations. It was
necessary to allow the representation of two parametrized types for
classes in the intermediate representation. Also, it was necessary that
the first and second compiler passes could recognized map and filter
operations. These functions are not being translated on the run-time
translation layer, though.
References #22 and #23.