Refactoring prior to filter and map impl in PM
Some necessary refactoring prior to filter and map implementation in
ParallelME runtime.

Also, I fixed a bug (that was not documented) that caused the runtime
kernel to thrown errors when the user code had references to Pixel x and
y coordinates due to the lack of these parameters in user function
kernels.

Referencing #22 and #23.